+++ Assets/_DB/Scripts/Game/GameManagers/Pause/IDBPauseCurrencyManager.cs
@@ -2,6 +2,6 @@ namespace DB_Game
 {
     public interface IDBPauseCurrencyManager
     {
-        int PassedTimeFoodRewardCalc(int timePassed);
+        double PassedTimeFoodRewardCalc(int timePassed);
     }
 }
